+++ openeo/udf/xarraydatacube.py
@@ -9,7 +9,7 @@ import collections
 import json
 import typing
 from pathlib import Path
-from typing import Union
+from typing import Union, Optional
 
 import numpy
 import xarray
@@ -314,9 +314,9 @@ class XarrayIO:
         return r.transpose(*dims)
 
     @classmethod
-    def from_netcdf_file(cls, path: Union[str, Path]) -> xarray.DataArray:
+    def from_netcdf_file(cls, path: Union[str, Path], engine: Optional[str] = None) -> xarray.DataArray:
         # load the dataset and convert to data array
-        ds = xarray.open_dataset(path, engine='h5netcdf')
+        ds = xarray.open_dataset(path, engine=engine)
 
         # Skip non-numerical variables (like "crs")
         band_vars = [k for k, v in ds.data_vars.items() if v.dtype.kind in {"b", "i", "u", "f"} and len(v.dims) > 0]
@@ -360,7 +360,7 @@ class XarrayIO:
             custom_print(jsonarray)
 
     @classmethod
-    def to_netcdf_file(cls, array: xarray.DataArray, path: Union[str, Path]):
+    def to_netcdf_file(cls, array: xarray.DataArray, path: Union[str, Path], engine: Optional[str] = None):
         # temp reference to avoid modifying the original array
         result = array
         # rearrange in a basic way because older xarray versions have a bug and ellipsis don't work in xarray.transpose()
@@ -375,5 +375,4 @@ class XarrayIO:
                 labels = ['band_' + str(i) for i in range(result.shape[result.dims.index('bands')])]
                 result = result.assign_coords(bands=labels)
         result = result.to_dataset('bands')
-        result.to_netcdf(path, engine='h5netcdf')
-
+        result.to_netcdf(path, engine=engine)